pyproject.toml
[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310']

[project]
name = "cat_breeds"
version = "0.1.0"
description = "Cat breed identification and querying using GenAI and image embeddings"
authors = [
    { name="Gaby Agrocostea", email="gabya06@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"

dependencies = [
    "chromadb>=1.0.3,<2.0.0",
    "google-genai>=1.7.0,<2.0.0",
    "google-api-core>=2.24.2",
    "matplotlib==3.8.0",
    "numpy>=1.26.0,<2.0.0",
    "openai==0.27.2",
    "pandas==2.2.3",
    "Pillow==9.4.0",
    "requests==2.31.0",
    "streamlit>=1.44.1",
    "torch==2.6.0+cpu ; platform_system != 'Darwin'",
    "torch==2.6.0 ; platform_system == 'Darwin'",  # macOS doesn't use `+cpu` tags    
    "tqdm==4.65.0",
    "transformers==4.45.2",
    "watchdog>=4.0.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"


# Tell setuptools to look in src/ for packages
[tool.setuptools.packages.find]
where = ["src"]